+++ lib/attributes_and_token_lists/engine.rb
@@ -16,6 +16,7 @@ module AttributesAndTokenLists
         ApplicationController.renderer.render(template: "initializers/attributes_and_token_lists")
       rescue ActionView::MissingTemplate
       end
+      ActiveSupport.run_load_hooks :attributes_and_token_lists, AttributesAndTokenLists
 
       AttributesAndTokenLists.define_builder_helper_methods(ActionView::Base)
     end